v0.1.0
This commit contains some very harsh changes made for the first tagged
version.

This version contains changes made to use the repository as a library.

In addition to the library, these changes maintain the ability to download
and compile a binary to play the tic tac toe game.

This change breaks backwards compatibility with previously created pseudo
versions.

The module suffix was also renamed from tic-tac-toe to xogo.